+++ src/main/java/nl/kb/dare/oai/GetRecordOperations.java
@@ -5,9 +5,9 @@ import nl.kb.stream.ByteCountOutputStream;
 import nl.kb.stream.ChecksumOutputStream;
 import nl.kb.filestorage.FileStorage;
 import nl.kb.filestorage.FileStorageHandle;
-import nl.kb.dare.http.HttpFetcher;
-import nl.kb.dare.http.HttpResponseHandler;
-import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
+import nl.kb.http.HttpFetcher;
+import nl.kb.http.HttpResponseHandler;
+import nl.kb.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.manifest.ManifestFinalizer;
 import nl.kb.dare.manifest.ManifestXmlHandler;
 import nl.kb.dare.manifest.ObjectResource;
