+++ src/main/java/nl/kb/http/HttpFetcher.java
@@ -1,4 +1,4 @@
-package nl.kb.dare.http;
+package nl.kb.http;
 
 import java.net.URL;
