+++ src/test/java/nl/kb/dare/endpoints/OaiRecordsEndpointTest.java
@@ -2,8 +2,8 @@ package nl.kb.dare.endpoints;
 
 import com.google.common.collect.Lists;
 import nl.kb.filestorage.FileStorage;
-import nl.kb.dare.http.HttpFetcher;
-import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
+import nl.kb.http.HttpFetcher;
+import nl.kb.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.model.oai.OaiRecord;
 import nl.kb.dare.model.oai.OaiRecordDao;
 import nl.kb.dare.model.oai.OaiRecordQuery;
