+++ src/test/java/nl/kb/dare/model/repository/RepositoryValidatorTest.java
@@ -1,10 +1,10 @@
 package nl.kb.dare.model.repository;
 
 
-import nl.kb.dare.http.HttpFetcher;
-import nl.kb.dare.http.HttpResponseException;
-import nl.kb.dare.http.HttpResponseHandler;
-import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
+import nl.kb.http.HttpFetcher;
+import nl.kb.http.HttpResponseException;
+import nl.kb.http.HttpResponseHandler;
+import nl.kb.http.responsehandlers.ResponseHandlerFactory;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
