+++ src/test/java/nl/kb/dare/oai/ScheduledOaiHarvesterTest.java
@@ -3,7 +3,7 @@ package nl.kb.dare.oai;
 import com.google.common.collect.Lists;
 import nl.kb.filestorage.FileStorage;
 import nl.kb.filestorage.FileStorageHandle;
-import nl.kb.dare.http.responsehandlers.ResponseHandlerFactory;
+import nl.kb.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.model.oai.OaiRecord;
 import nl.kb.dare.model.oai.OaiRecordDao;
 import nl.kb.dare.model.reporting.ErrorReportDao;
